refactor(logger): move logger instance and config to logging.py
